backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
billiard==3.3.0.23        # via celery
celery==3.1.23
django-celery==3.1.17
django==1.10
djangorestframework==3.4.1
kombu==3.0.35             # via celery
psycopg2==2.6.2
pydub==0.16.5
pytz==2016.6.1            # via celery
redis==2.10.5
